:art: Improve code structure
